console: move struct lxc_terminal{_info}
